map-tools re-import and a bunch of grunt task work (#3)
This PR:
- re-imports the map-tools package
- grunt task work in jasmine, eslint, babel, sass, e2e, & watch
- cleans up a bunch of package-level stuff that was duplicated at the root

I'm not sure why Travis integration is broken for this repo. I've sent an email to travis support to look into it.

Finally, I'm still not 100% sure that this will scale well with very many packages but I'm hopeful. It sure seems like an easier problem to tackle as opposed to maintaining multiple separate projects.